requirements.txt
Django==4.1.7
djangorestframework==3.14.0
pytest==7.2.2
requests==2.28.2
structlog==22.3.0
